pyproject.toml
[tool.poetry]
name = "airlift"
version = "0.1.0"
description = "A CLI for creating a flexible Apache Airflow local development environment"
authors = ["jered.little <jeredlittle1996@gmail.com>"]
readme = "README.md"
packages = [{ include = "airlift", from = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
argparse = "^1.4.0"
halo = "^0.0.31"
pyyaml = "^6.0"
hiyapyco = "^0.5.1"
jinja2 = "^3.1.2"
dotmap = "^1.3.30"
requests = "^2.30.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
coverage = "^7.2.5"
mock = "^5.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
airlift = "airlift.main:main"
